api/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.4
build==1.0.3
CacheControl==0.13.1
certifi==2024.6.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
command-not-found==0.3
crashtest==0.4.1
cryptography==3.4.8
dbus-python==1.2.18
Deprecated==1.2.14
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.1
django-oauth-toolkit==2.1.0
django-redis==5.2.0
dnspython==2.6.1
dulwich==0.21.7
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
filelock==3.13.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib-metadata==7.0.1
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.3.1
jeepney==0.7.1
Jinja2==3.1.4
jwcrypto==1.5.1
keyring==24.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==8.10.0
msal==1.29.0
msgpack==1.0.7
netifaces==0.11.0
oauthlib==3.2.0
onetimepass==1.0.1
orjson==3.10.5
outcome==1.3.0.post0
packaging==23.2
pdfkit==1.0.0
pexpect==4.9.0
Pillow==9.2.0
pkginfo==1.9.6
platformdirs==3.11.0
ply==3.11
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
psycopg2==2.9.9
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic-extra-types==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.18.4
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.1.0
pyparsing==2.4.7
pyproject_hooks==1.0.0
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-ldap==3.4.4
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
redis==5.0.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
SecretStorage==3.3.1
selenium==4.22.0
shellingham==1.5.4
six==1.16.0
sloth-framework==0.1.16
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
sqlparse==0.4.4
ssh-import-id==5.11
starlette==0.37.2
systemd-python==234
tomli==2.0.1
tomlkit==0.12.3
trio==0.25.1
trio-websocket==0.11.1
trove-classifiers==2024.2.22
typer==0.12.3
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.25.1
wadllib==1.3.6
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
xlrd==1.2.0
xlwt==1.3.0
zipp==1.0.0
